requirements.txt
moviepy
numpy
openai-whisper
customtkinter
pillow
ffmpeg-python
tqdm
opencv-python
colorama
pydub
